requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.0
beautifulsoup4==4.12.3
pydantic==2.9.2
google-generativeai==0.3.2
pytest==7.4.0
pytest-asyncio==0.21.0
responses==0.23.0
playwright==1.40.0